+++ refinedstorage-platform-api/src/main/java/com/refinedmods/refinedstorage/platform/api/support/resource/ResourceContainerInsertStrategy.java
@@ -21,7 +21,7 @@ public interface ResourceContainerInsertStrategy {
      */
     Optional<InsertResult> insert(ItemStack container, ResourceAmount resourceAmount);
 
-    Optional<ConversionInfo> getConversionInfo(ResourceKey resource);
+    Optional<ConversionInfo> getConversionInfo(ResourceKey resource, ItemStack container);
 
     record InsertResult(ItemStack container, long inserted) {
     }