+++ refinedstorage-platform-common/src/main/java/com/refinedmods/refinedstorage/platform/common/grid/screen/AbstractGridScreen.java
@@ -312,7 +312,7 @@ public abstract class AbstractGridScreen<T extends AbstractGridContainerMenu> ex
             addDetailedTooltip(view, platformResource, processedLines);
         }
         if (!platformResource.isZeroed()) {
-            processedLines.addAll(platformResource.getExtractionHints());
+            processedLines.addAll(platformResource.getExtractionHints(getMenu().getCarried()));
         }
         Platform.INSTANCE.renderTooltip(graphics, processedLines, mouseX, mouseY);
     }
@@ -385,7 +385,7 @@ public abstract class AbstractGridScreen<T extends AbstractGridContainerMenu> ex
         final ItemStack carriedStack = getMenu().getCarried();
         final PlatformGridResource resource = getCurrentGridResource();
 
-        if (resource != null && carriedStack.isEmpty()) {
+        if (resource != null) {
             mouseClickedInGrid(clickedButton, resource);
             return true;
         }